source/services/ComponentRegistry/dependent-api-discovery/requirements.txt
python-dotenv  # >=1.0.0